Issues-14: Moved TickerActivity to separate module.
