+++ app/build.gradle
@@ -31,8 +31,6 @@ dependencies {
 
     implementation 'com.android.support:appcompat-v7:28.0.0'
 
-    implementation 'com.robinhood.ticker:ticker:2.0.1'
-
     //Particle View
     implementation 'me.wangyuwei:ParticleView:1.0.4'
 
@@ -53,6 +51,7 @@ dependencies {
 
     implementation project(path: ':alerter')
     implementation project(path: ':loadingbutton')
+    implementation project(path: ':ticker')
 }
 
 // For resolve library conflict(dou navigation drawer)
