+++ app/src/main/AndroidManifest.xml
@@ -16,8 +16,6 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".ticker.TickerActivity" />
-
         <activity android:name=".particle.ParticleActivity" />
 
         <activity
