+++ app/src/main/java/com/kirchhoff/animation/MainActivity.java
@@ -5,6 +5,7 @@ import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 
+import com.example.TickerActivity;
 import com.example.alerter.AlerterActivity;
 import com.example.loadingbutton.LoadingButtonActivity;
 import com.kirchhoff.animation.expandable.ExpandableLinearLayoutActivity;
@@ -13,7 +14,6 @@ import com.kirchhoff.animation.recycler.focusresize.FocusResizeActivity;
 import com.kirchhoff.animation.recycler.vega.VegaManagerActivity;
 import com.kirchhoff.animation.scaling.ScalingMainActivity;
 import com.kirchhoff.animation.swipeback.SwipeBackLayoutActivity;
-import com.kirchhoff.animation.ticker.TickerActivity;
 
 /**
  * @author Kirchhoff-
