+++ app/src/main/res/layout/activity_main.xml
@@ -19,7 +19,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:onClick="ticker"
-            android:text="@string/ticker.library" />
+            android:text="@string/ticker" />
 
         <Button
             android:id="@+id/interpolators"
