+++ ticker/src/main/AndroidManifest.xml
@@ -1,0 +1,7 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.example.ticker">
+
+    <application>
+        <activity android:name="com.example.TickerActivity" />
+    </application>
+</manifest>
